Refactor internal actions: move ElevenLabs functionality to createNode and remove eleven.ts
